# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A e-commerce API server",
        "title": "ECOM API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/orders/{id}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理員更新訂單狀態，已完成或已取消的訂單不可更新(僅管理員)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "更新訂單狀態",
                "parameters": [
                    {
                        "type": "string",
                        "description": "訂單 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "新狀態",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateOrderStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.OrderResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "請求參數錯誤或訂單狀態不允許更新",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    },
                    "401": {
                        "description": "未授權",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    },
                    "403": {
                        "description": "無管理員權限",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    },
                    "404": {
                        "description": "訂單不存在",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "伺服器錯誤",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    }
                }
            }
        },
...
        "/orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "取得當前使用者的所有訂單列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "取得用戶所有訂單",
                "responses": {
                    "200": {
                        "description": "取得成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.OrderResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授權",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "伺服器錯誤",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "將購物車中的商品轉換為訂單（結帳）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "從購物車建立訂單",
                "parameters": [
                    {
                        "description": "訂單資訊",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "訂單建立成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.OrderResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "請求參數錯誤",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    },
                    "401": {
                        "description": "未授權",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    },
                    "404": {
                        "description": "購物車為空或商品不存在",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    },
                    "409": {
                        "description": "庫存不足",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "伺服器錯誤",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "取得指定訂單的詳細資訊",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "取得訂單詳情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "訂單 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "取得成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.OrderResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授權",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    },
                    "403": {
                        "description": "無權限訪問此訂單",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    },
                    "404": {
                        "description": "訂單不存在",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "伺服器錯誤",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    }
                }
            }
        },
        "/orders/{id}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "取消指定的訂單（僅限 Pending 或 Processing 狀態）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "取消訂單",
                "parameters": [
                    {
                        "type": "string",
                        "description": "訂單 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "取消成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.OrderResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "訂單無法取消",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    },
                    "401": {
                        "description": "未授權",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    },
                    "403": {
                        "description": "無權限",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    },
                    "404": {
                        "description": "訂單不存在",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "伺服器錯誤",
                        "schema": {
                            "$ref": "#/definitions/dto.StandardResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateOrderRequest": {
            "type": "object",
            "required": [
                "recipient_name",
                "shipping_address"
            ],
            "properties": {
                "recipient_name": {
                    "type": "string",
                    "maxLength": 50
                },
                "shipping_address": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
...
        "dto.OrderItemResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "product": {
                    "$ref": "#/definitions/dto.ProductResponse"
                },
                "product_id": {
                    "type": "string"
                },
                "product_name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "subtotal": {
                    "type": "number"
                },
                "unit_price": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.OrderResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.OrderItemResponse"
                    }
                },
                "order_date": {
                    "type": "string"
                },
                "recipient_name": {
                    "type": "string"
                },
                "shipping_address": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.ProductResponse": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "category_name": {
                    "type": "string"
                },
                "category_slug": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "stock_quantity": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "dto.StandardResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.UpdateOrderStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string",
                    "enum": [
                        "Pending",
                        "Processing",
                        "Shipped",
                        "Completed",
                        "Cancelled"
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}